Create Email Slicer with Python
Finally, we are just formatting to print the output. The above code can be enhanced with more ideas depending on your needs. As a beginner, you must try these types of programs to improve your coding skills. In the long run, it will also help you build your algorithms and increase your ability to think logically.