Create COVID19 Vaccine Tracker Using Python
As we know the world is facing an unprecedented challenge with communities and economies everywhere affected by the COVID19. So, we are going to do some fun during this time by tracking their vaccine. Let’s see a simple Python script to improve for tracking the COVID19 vaccine.

Modules Needed
pip install bs4 
pip install request